[charan] setting java classes

+++ app/src/main/java/com/example/employeepayrollsystem/models/PartTime.java
@@ -2,17 +2,17 @@ package com.example.employeepayrollsystem.models;
 
 public class PartTime extends Employee {
 
-    private float rate;
+    private double rate;
     private float hoursWorked;
 
-    public PartTime(int id, String name, int birthYear, float rate, float hoursWorked, Vehicle vehicle) {
+    public PartTime(int id, String name, int birthYear, double rate, float hoursWorked, Vehicle vehicle) {
         super(id, name, birthYear, vehicle);
         this.rate = rate;
         this.hoursWorked =hoursWorked;
     }
 
 
-    public float getRate() { return rate; }
+    public double getRate() { return rate; }
 
     public void setRate(int rate) { this.rate = rate; }
